CI: link ws2_32 for the arcade bridge; surface MSVC diagnostics as annotations
The Windows build (x64/ARM64/ARM64EC) fails on every target that pulls in
ArcadeServer.cpp: ok_core never linked ws2_32 even though the new web bridge
calls Winsock (WSAStartup/socket/setsockopt/select/accept/send/recv), so
ok_arcade_server_tests and KieeKeyApp.exe stopped linking on MSVC. ws2_32 is
added to ok_core's interface libs (arcade_serve already had it).

Because the raw Actions log is not downloadable everywhere, the Build step now
tees the compiler output and re-emits the first 40 diagnostics as
::error:: annotations (readable from the check-run API) and uploads the full log
as an artifact on failure. That makes the MSVC-only errors visible without
guessing.

+++ src/core/ArcadeServer.cpp
@@ -263,7 +263,7 @@ std::size_t ArcadeServer::pendingResults() const {
 void ArcadeServer::drainRunResults() {
     // Shared with the desktop hub (ArcadeManager::drainRunResultsToProgression):
     // one implementation, so the two front-ends can never credit differently.
-    m_runCounter += m_manager.drainRunResultsToProgression();
+    m_runCounter += static_cast<std::uint32_t>(m_manager.drainRunResultsToProgression());
 }
 
 bool ArcadeServer::startGame(const std::string& slug) {
